worker-safety-service/pyproject.toml
[tool.poetry]
name = "worker-safety-service"
version = "1.12.18"
description = "Worker Safety Services"
authors = ["Urbint Engineering <engineering@urbint.com>"]

[tool.poetry.scripts]
wss = "worker_safety_service.cli.main:app"
seed = "support.scripts.seed.main:app"
dev = "support.scripts.dev.main:app"

[tool.poetry.dependencies]
alembic = "^1.13.1"
ddtrace = "^2.8.5"
dependency-injector = "^4.41.0"
fastapi = "^0.105.0"
google-cloud-storage = "^2.15.0"
openpyxl = "^3.1.2"
pendulum = "^3.0.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.1"
pydantic = "^1.10.13"
PyJWT = { extras = ["crypto"], version = "^2.8.0" }
python = "^3.10"
numpy = "<2.3.0"
sqlalchemy2-stubs = "^0.0.2a38"
sqlmodel = "^0.0.11"
strawberry-graphql = { extras = ["fastapi"], version = "^0.219.2" }
structlog = "^24.1.0"
typer = "^0.9.0"
uvicorn = { extras = ["standard"], version = "^0.25.0" }
jsonpatch = "^1.33"
redis = "^5.0.1"
pyfailsafe = "^0.6.0"
httpx = "^0.27.0"
Brotli = "^1.1.0"
shapely = "^1.8.4"
tomli = "^2.0.1"
wrapt = "^1.16.0"
mmh3 = "^4.1.0"
starlette = "^0.27.0"
launchdarkly-server-sdk = "^9.1.0"

google-auth = "2.29.0"
protobuf = "^5.28.2"
firebase-admin = "^6.5.0"
python-multipart = "^0.0.20"
pandas = "^2.3.1"
xlrd = "^2.0.2"
[tool.poetry.dev-dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.7"
pydantic-factories = "^1.17.3"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
pytest-timeout = "^2.2.0"
pytest-xdist = "^3.5.0"
colorama = "^0.4.6"
tqdm = "^4.66.4"
types-certifi = "^2021.10.8"

[tool.poetry.group.dev.dependencies]
pytest-rerunfailures = "^14.0"
types-redis = "^4.6.0.11"
types-pyyaml = "^6.0.12.12"
autohooks-plugin-black = "^23.10.0"
autohooks-plugin-mypy = "^23.10.0"
autohooks-plugin-isort = "^23.10.0"
autohooks-plugin-flake8 = "^23.10.0"
types-pytz = "^2024.1.0.20240417"
pandas-stubs = "^2.3.0.250703"


[tool.pytest.ini_options]
markers = ["gcs", "integration", "unit"]
timeout = 20
timeout_method = 'thread'
asyncio_mode = 'auto'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autohooks]
mode = "poetry"
pre-commit = [
  "autohooks.plugins.black",
  "autohooks.plugins.flake8",
  "autohooks.plugins.isort",
  "autohooks.plugins.mypy",
]


[tool.isort]
profile = "black"

[tool.autohooks.plugins.flake8]
arguments = ["--config=setup.cfg"]

[tool.autohooks.plugins.mypy]
arguments = ["--config-file=setup.cfg"]
